+++ packages/graph-editor/package.json
@@ -40,6 +40,7 @@
     "@rematch/persist": "2.1.2",
     "@stitches/react": "1.2.8",
     "@tokens-studio/graph-engine": "*",
+    "@tokens-studio/sdk": "^0.3.0",
     "@tokens-studio/tokens": "0.0.21",
     "@tokens-studio/types": "0.2.3",
     "@tokens-studio/ui": "0.3.0",
@@ -55,7 +56,9 @@
     "json5": "2.2.3",
     "jszip": "3.10.1",
     "postcss-value-parser": "4.2.0",
+    "react": "18.2.0",
     "react-color": "2.19.3",
+    "react-dom": "18.2.0",
     "react-drag-reorder": "1.2.0",
     "react-error-boundary": "4.0.4",
     "react-folder-tree": "5.1.1",
